test(mocks): add postgres mock for testing SQL client types
Refactor test mocking approach by extracting postgres mock implementation into a separate module.
Simplifies type testing for PgflowSqlClient by moving complex mock setup to a reusable function.
