+++ pkgs/core/__tests__/mocks/index.ts
@@ -1,0 +1,1 @@
+export * from './postgres.ts';
